Feature(react-menu-grid-preview): Add grouping items story
