+++ packages/react-components/react-menu-grid-preview/stories/src/MenuGrid/MenuGridDefault.stories.tsx
@@ -1,17 +1,8 @@
 import * as React from 'react';
 import { Button, Menu, MenuPopover, MenuTrigger } from '@fluentui/react-components';
-import {
-  MenuGrid,
-  MenuGridCell,
-  MenuGridRow,
-  MenuGridRowGroup,
-  MenuGridRowGroupHeader,
-} from '@fluentui/react-menu-grid-preview';
+import { MenuGrid, MenuGridCell, MenuGridRow } from '@fluentui/react-menu-grid-preview';
 
-const items = {
-  people: ['Olivia Carter', 'Liam Thompson', 'Sophia Martinez', 'Noah Patel', 'Emma Robinson'],
-  agentsAndBots: ['Facilitator', 'Copilot'],
-};
+const items = ['Olivia Carter', 'Liam Thompson', 'Sophia Martinez', 'Noah Patel', 'Emma Robinson'];
 
 export const Default = () => {
   return (
@@ -21,34 +12,17 @@ export const Default = () => {
       </MenuTrigger>
       <MenuPopover>
         <MenuGrid>
-          <MenuGridRowGroup>
-            <MenuGridRowGroupHeader>People</MenuGridRowGroupHeader>
-            {items.people.map((name, index) => (
-              <MenuGridRow key={index} aria-label={name}>
-                <MenuGridCell>{name}</MenuGridCell>
-                <MenuGridCell>
-                  <Button aria-label={`Profile card for ${name}`}>Avatar icon</Button>
-                </MenuGridCell>
-                <MenuGridCell>
-                  <Button>Remove {name}</Button>
-                </MenuGridCell>
-              </MenuGridRow>
-            ))}
-          </MenuGridRowGroup>
-          <MenuGridRowGroup>
-            <MenuGridRowGroupHeader>Agents and bots</MenuGridRowGroupHeader>
-            {items.agentsAndBots.map((name, index) => (
-              <MenuGridRow key={index} aria-label={name}>
-                <MenuGridCell>{name}</MenuGridCell>
-                <MenuGridCell>
-                  <Button aria-label={`Profile card for ${name}`}>Avatar icon</Button>
-                </MenuGridCell>
-                <MenuGridCell>
-                  <Button>Remove {name}</Button>
-                </MenuGridCell>
-              </MenuGridRow>
-            ))}
-          </MenuGridRowGroup>
+          {items.map((name, index) => (
+            <MenuGridRow key={index} aria-label={name}>
+              <MenuGridCell>{name}</MenuGridCell>
+              <MenuGridCell>
+                <Button aria-label={`Profile card for ${name}`}>Avatar icon</Button>
+              </MenuGridCell>
+              <MenuGridCell>
+                <Button>Remove {name}</Button>
+              </MenuGridCell>
+            </MenuGridRow>
+          ))}
         </MenuGrid>
       </MenuPopover>
     </Menu>
