+++ packages/react-components/react-menu-grid-preview/stories/src/MenuGrid/MenuGridDescription.md
@@ -1,1 +1,1 @@
-A menu grid displays a complex menu structured as grid with more actions. It is usually rendered inside of the Menu component.
+A menu grid displays a complex menu structured as grid with more actions. It is usually rendered inside of the `Menu` component.
