+++ packages/react-components/react-menu-grid-preview/stories/src/MenuGrid/index.stories.tsx
@@ -8,6 +8,7 @@ import {
 import descriptionMd from './MenuGridDescription.md';
 
 export { Default } from './MenuGridDefault.stories';
+export { GroupingItems } from './MenuGridGroupingItems.stories';
 export { WithSubmenu } from './MenuGridWithSubmenu.stories';
 
 export default {
